reworked the way the bot counted words
The bot used to count words using js and was built into the bot. This 
caused the bot to seize up upon recieving many messages as it had to 
process all of the spam before it could respond to commands. This made 
the bot slow. Now the bot should execute a python script that should 
process the messages on a different thread