added a testing command and a template command
The testing command was for testing python scripts ran through the js 
code. The template command is there to make it easier to make new 
commands. both are unable to executed through the chat because their 
names are capitalized which cannot be achieved due to how the bot 
processes commands